// src/app/Aqar/aqarReportPage/models/aqarReport.ts
// tslint:disable-next-line:class-name
export class aqarReport{
    'id': number;
    'city': string;
    'district': string;
    'suk_number': string;
    'suk_date': string;
    'evaluation_date': string;
    'street_asphalt_type': number;
    'building_area': string;
    'land_area': string;
    'interfaces_number': number;
    'street_type': number;
    'piece_number': string;
    'no_planned': string;
    'aqar_type': number;
    'aqar_age': string;
    'building_status': number;
    'unit_number_apartments': string;
    'telephone_service': number;
    'electricity_service': number;
    'sanitation_service': number;
    'water_service': number;
    'building_meter_price': number;
    'land_meter_price': number;
    'land_price': number;
    'building_price': number;
    'area': number;
    'meter_price': number;
    'price': number;
}